move the `write_km_eroded_volumes` to the `bank_erosion.utils`
