+++ src/dfastbe/bank_erosion/bank_erosion.py
@@ -54,11 +54,10 @@ from dfastbe.bank_erosion.data_models import (
     WaterLevelData,
 )
 from dfastbe.bank_erosion.debugger import Debugger
-from dfastbe.bank_erosion.utils import BankLinesProcessor, intersect_line_mesh
+from dfastbe.bank_erosion.utils import BankLinesProcessor, intersect_line_mesh, write_km_eroded_volumes
 from dfastbe.io.config import (
     ConfigFile,
     log_text,
-    write_km_eroded_volumes,
 )
 from dfastbe.io.data_models import LineGeometry
 from dfastbe.kernel import get_zoom_extends
